requirements.txt
fastapi
uvicorn
numpy
pydantic
python-dotenv
requests
